feat: support SearchPanel component export
