+++ .changeset/angry-berries-live.md
@@ -1,0 +1,7 @@
+---
+'@rspress/core': patch
+---
+
+feat: support SearchPanel component export
+
+feat: 支持 SearchPanel 组件导出
